Refactor: Create FormOne and FormTwo components to clean up App.jsx
