+++ 12.webDeveloper/example/src/main.jsx
@@ -1,10 +1,10 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import App, { App2 } from './App.jsx'
+import App from './App.jsx'
 
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <App2 />
+    <App />
   </StrictMode>,
 )
